refactor: sidebar nav folders
